+++ src/Container.php
@@ -16,6 +16,7 @@ use Castor\Http\HttpDownloader;
 use Castor\Import\Importer;
 use Castor\Import\Mounter;
 use Castor\Import\Remote\Composer;
+use Castor\Runner\LlmRunner;
 use Castor\Runner\ParallelRunner;
 use Castor\Runner\PhpRunner;
 use Castor\Runner\ProcessRunner;
@@ -53,6 +54,7 @@ final class Container
         public readonly Importer $importer,
         public readonly InputInterface $input,
         public readonly Kernel $kernel,
+        public readonly LlmRunner $llmRunner,
         public readonly LoggerInterface $logger,
         public readonly Mounter $mounter,
         public readonly Notifier $notifier,